Updated profile layout
