+++ apps/admin-x-activitypub/src/views/Profile/components/ProfilePage.tsx
@@ -1,11 +1,14 @@
 import APAvatar from '@src/components/global/APAvatar';
+import FollowButton from '@src/components/global/FollowButton';
 import Layout from '@src/components/layout';
 import {Account} from '@src/api/activitypub';
-import {Button, Heading, Tab, TabView} from '@tryghost/admin-x-design-system';
+import {Button, Heading, Icon, LoadingIndicator, NoValueLabel, Tab, TabView} from '@tryghost/admin-x-design-system';
 import {ProfileTab} from '../Profile';
 import {Skeleton} from '@tryghost/shade';
 import {useEffect, useRef, useState} from 'react';
 import {useParams} from '@tryghost/admin-x-framework';
+
+const noop = () => {};
 
 type ProfilePageProps = {
     account: Account,
@@ -31,6 +34,12 @@ const ProfilePage:React.FC<ProfilePageProps> = ({
 }) => {
     const [selectedTab, setSelectedTab] = useState<ProfileTab>('posts');
     const params = useParams();
+
+    useEffect(() => {
+        setSelectedTab('posts');
+    }, [params.handle]);
+
+    const isCurrentUser = params.handle === account?.handle || !params.handle;
 
     const tabs = [
         {
@@ -75,75 +84,94 @@ const ProfilePage:React.FC<ProfilePageProps> = ({
 
     return (
         <Layout>
-            <div className='relative isolate'>
-                <div className='absolute -right-8 left-0 top-0 z-0 h-[5vw]'>
-                    {account?.bannerImageUrl &&
-                    <div className='size-full'>
-                        <img
-                            alt={account?.name}
-                            className='size-full object-cover'
-                            src={account?.bannerImageUrl}
-                        />
-                    </div>
-                    }
-                </div>
-                <div className='relative z-10 mx-auto flex w-full max-w-[620px] flex-col items-center pb-16 pt-[calc(5vw-52px)]'>
-                    <div className='mx-auto w-full'>
-                        <div>
-                            <div className='flex items-end justify-between'>
-                                <div className='-ml-2 rounded-full bg-white p-1 dark:bg-black dark:outline-black'>
-                                    <APAvatar
-                                        author={account && {
-                                            icon: {
-                                                url: account?.avatarUrl
-                                            },
-                                            name: account?.name,
-                                            handle: account?.handle
-                                        }}
-                                        size='lg'
+            <div className='z-0 -mx-8 -mt-9 flex flex-col items-center pb-16'>
+                <div className='mx-auto w-full'>
+                    {isLoadingAccount && (
+                        <LoadingIndicator size='lg' />
+                    )}
+                    {!isLoadingAccount && !account && (
+                        <NoValueLabel icon='user-add'>
+                        Profile not found
+                        </NoValueLabel>
+                    )}
+                    {!isLoadingAccount && account && (
+                        <>
+                            {account.bannerImageUrl ?
+                                <div className='h-[15vw] w-full overflow-hidden bg-gradient-to-tr from-gray-200 to-gray-100'>
+                                    <img
+                                        alt={account.name}
+                                        className='size-full object-cover'
+                                        src={account.bannerImageUrl}
                                     />
                                 </div>
-                            </div>
-                            <Heading className='mt-4' level={3}>{!isLoadingAccount ? account?.name : <Skeleton className='w-32' />}</Heading>
-                            <span className='mt-1 text-[1.5rem] text-gray-700 dark:text-gray-600'>
-                                <span>{!isLoadingAccount ? account?.handle : <Skeleton className='w-full max-w-56' />}</span>
-                            </span>
-                            {(account?.bio || customFields.length > 0 || isLoadingAccount) && (
-                                <div ref={contentRef} className={`ap-profile-content transition-max-height relative text-[1.5rem] duration-300 ease-in-out [&>p]:mb-3 ${isExpanded ? 'max-h-none pb-7' : 'max-h-[160px] overflow-hidden'} relative`}>
-                                    <div className='ap-profile-content mt-3 text-[1.5rem] [&>p]:mb-3'>
-                                        {!isLoadingAccount ?
-                                            <div dangerouslySetInnerHTML={{__html: account?.bio ?? ''}} /> :
-                                            <>
-                                                <Skeleton />
-                                                <Skeleton className='w-full max-w-48' />
-                                            </>
-                                        }
+                                :
+                                <div className='h-[8vw] w-full overflow-hidden bg-gradient-to-tr from-white to-white'></div>
+                            }
+                            <div className='mx-auto -mt-12 max-w-[620px] px-6'>
+                                <div className='flex items-end justify-between'>
+                                    <div className='-ml-2 rounded-full bg-white p-1 dark:bg-black'>
+                                        <APAvatar
+                                            author={
+                                                {
+                                                    icon: {
+                                                        url: account.avatarUrl
+                                                    },
+                                                    name: account.name,
+                                                    handle: account.handle
+                                                }
+                                            }
+                                            size='lg'
+                                        />
                                     </div>
-                                    {customFields.map(customField => (
-                                        <span key={customField.name} className='mt-3 line-clamp-1 flex flex-col text-[1.5rem]'>
-                                            <span className={`text-xs font-semibold`}>{customField.name}</span>
-                                            <span dangerouslySetInnerHTML={{__html: customField.value}} className='ap-profile-content truncate'/>
+                                    {!isCurrentUser &&
+                                    <FollowButton
+                                        following={account.followedByMe}
+                                        handle={account.handle}
+                                        type='primary'
+                                        onFollow={noop}
+                                        onUnfollow={noop}
+                                    />
+                                    }
+                                </div>
+                                <Heading className='mt-4' level={3}>{!isLoadingAccount ? account.name : <Skeleton className='w-32' />}</Heading>
+                                <a className='group/handle mt-1 flex items-center gap-1 text-[1.5rem] text-gray-800 hover:text-gray-900' href={account.url} rel='noopener noreferrer' target='_blank'>
+                                    <span>{!isLoadingAccount ? account?.handle : <Skeleton className='w-full max-w-56' />}</span>
+                                    <Icon className='opacity-0 transition-opacity group-hover/handle:opacity-100' name='arrow-top-right' size='xs'/>
+                                </a>
+                                {(account.bio || customFields.length > 0) && (<div ref={contentRef} className={`ap-profile-content transition-max-height relative text-[1.5rem] duration-300 ease-in-out [&>p]:mb-3 ${isExpanded ? 'max-h-none pb-7' : 'max-h-[160px] overflow-hidden'} relative`}>
+                                    {!isLoadingAccount ?
+                                        <div dangerouslySetInnerHTML={{__html: account?.bio ?? ''}} /> :
+                                        <>
+                                            <Skeleton />
+                                            <Skeleton className='w-full max-w-48' />
+                                        </>
+                                    }
+                                    {customFields.map((attachment: {name: string, value: string}) => (
+                                        <span className='mt-3 line-clamp-1 flex flex-col text-[1.5rem]'>
+                                            <span className={`text-xs font-semibold`}>{attachment.name}</span>
+                                            <span dangerouslySetInnerHTML={{__html: attachment.value}} className='ap-profile-content truncate'/>
                                         </span>
                                     ))}
                                     {!isExpanded && isOverflowing && (
                                         <div className='absolute inset-x-0 bottom-0 h-16 bg-gradient-to-t from-white via-white/90 via-60% to-transparent' />
                                     )}
                                     {isOverflowing && <Button
-                                        className='absolute bottom-0 text-pink'
+                                        className='absolute bottom-0'
                                         label={isExpanded ? 'Show less' : 'Show all'}
                                         link={true}
+                                        size='sm'
                                         onClick={toggleExpand}
                                     />}
-                                </div>
-                            )}
-                            <TabView<ProfileTab>
-                                containerClassName='mt-6'
-                                selectedTab={selectedTab}
-                                tabs={tabs}
-                                onTabChange={setSelectedTab}
-                            />
-                        </div>
-                    </div>
+                                </div>)}
+                                <TabView<ProfileTab>
+                                    containerClassName='mt-6'
+                                    selectedTab={selectedTab}
+                                    tabs={tabs}
+                                    onTabChange={setSelectedTab}
+                                />
+                            </div>
+                        </>
+                    )}
                 </div>
             </div>
         </Layout>
